requirements.txt
streamlit
yfinance
pandas
plotly
gspread
google-auth
